+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 2.21.1 (2026-03-03)
+
+### Fix
+
+- **keyboard**: send correct key, code and keycode in type_text
+- **elements**: fix humanized interactions inside iframes
+- humanized scroll overshoot correction causes infinite loop
+
+## 2.21.0 (2026-03-01)
+
+### Feat
+
+- **interactions**: change humanize default from True to False
+
+### Fix
+
+- **elements**: forward humanize flag to click in type_text
+
 ## 2.20.2 (2026-02-18)
 
 ### Fix